+++ python-package/xgboost/VERSION
@@ -1,0 +1,1 @@
+0.4
